Set scope for remaining placement policy rules
This just sets the scope type for the remaining placement
policy rules that didn't have their scope type set when
the policy rule was created.

These all use system scope since that makes the most sense
for resources in Placement today. In the future, we might
have use cases where project admins can work with project-scoped
resources like allocations (and usages) but for now we just
consider everything owned by the system and scoped to the system.

By default, scope is not enforced in oslo.policy.

Part of blueprint granular-placement-policy

Change-Id: If86b808edac94275f0a650f7215001c39641eb0c