fix tox python3 overrides
We want to default to running all tox environments under python 3, so
set the basepython value in [testenv].

We do not want to specify a minor version number, because we do not
want to have to update the file every time we upgrade python.

We must explicitly set basepython in envs that are usually "automatic" -
otherwise due to https://github.com/tox-dev/tox/issues/425 these will
use `basepython` instead of their automatic equivalents.

Change-Id: I1026110f0f1736127cf0af8ec2ef791a0f999e3a
Signed-off-by: Doug Hellmann <doug@doughellmann.com>