Do not use nova.test in placement.test_deploy
The tests are very straightforward and do not require many of the setup
and fixtures that nova.test.NoDBTestCase supplies. Also, in the future
of placement extraction we don't want the dependecy.

One missing tool from NoDBTestCase is self.flags, which helps manage
configuration. That is a facade over CONF.set_override, so we just call
that instead.

Partially implements blueprint placement-extract

Change-Id: Ice6a42b3b8b94e155d7332256e9b568f80bc663d