Ignore UserWarning for scope checks during test runs
Placement API policy rules are defaulting to system scope.
Scope checks are disabled by default in oslo.policy, but
if you hit the API with a token that doesn't match the scope,
it generates a UserWarning, for every policy check on that
request. This is pretty annoying, so just filter those warnings
during our test runs.

Change-Id: I30ed00a96390d2c76cfc2a40c5a47c16eb51711c