Use placement context in placement functional tests
A few placement functional tests were importing and using
nova.context for a RequestContext. Placement has its own
in nova.api.openstack.placement.context so use that instead.

In the process, stop using an 'admin' context. It is not
necessary in placement as we do not do any policy checking
at the database level.

Change-Id: I1f6e6db6aca7dd160d3e94bb0b6ebf9b4f8dfd7d